Major updates
-Create combineInventoriesforFacilityList()
-Move more shared code from two combineInventories functions into addBaseInventoryIDs()
-Revise getInventoriesforFacilityMatches().Fixes issue #3
-Move colstokeep, addBaseInventoryIDs, getInventoriesforFacilityMatches, addChemicalMatches to globals
-Add docstrings to four remaining functions
